refactor(writer): rename faceutil to writer
fixes #2
