+++ honeybee_energy/_extend_honeybee.py
@@ -1,7 +1,7 @@
 from honeybee.properties import Properties
 import honeybee.writer as writer
 from .properties import EnergyProperties
-from .faceutil import face_to_idf
+from .writer import face_to_idf
 
 # add energy properties to Properties class
 Properties.energy =  property(lambda self: EnergyProperties(self.face_type))
